`Synthesizer`の構造改革をする
