+++ crates/voicevox_core/src/__internal.rs
@@ -1,4 +1,5 @@
 pub mod doctest_fixtures;
+pub mod interp;
 
 // VOICEVOX CORE内のラッパー向けの実装
 // FIXME: 要議論: https://github.com/VOICEVOX/voicevox_core/issues/595
