+++ crates/voicevox_core/src/engine/mod.rs
@@ -4,7 +4,6 @@ mod kana_parser;
 mod model;
 mod mora_list;
 mod open_jtalk;
-mod synthesis_engine;
 
 use super::*;
 
@@ -12,5 +11,5 @@ pub use self::acoustic_feature_extractor::*;
 pub use self::full_context_label::*;
 pub use self::kana_parser::*;
 pub use self::model::*;
+pub(crate) use self::mora_list::mora2text;
 pub use self::open_jtalk::OpenJtalk;
-pub use self::synthesis_engine::*;
