+++ crates/voicevox_core_c_api/Cargo.toml
@@ -24,6 +24,7 @@ derive-getters.workspace = true
 futures.workspace = true
 itertools.workspace = true
 libc = "0.2.134"
+ndarray.workspace = true
 once_cell.workspace = true
 serde_json.workspace = true
 thiserror.workspace = true
@@ -52,7 +53,6 @@ easy-ext.workspace = true
 inventory = "0.3.4"
 libloading = "0.7.3"
 libtest-mimic = "0.6.0"
-ndarray.workspace = true
 ndarray-stats = "0.5.1"
 regex.workspace = true
 serde.workspace = true
