+++ crates/voicevox_core_c_api/src/compatible_engine.rs
@@ -1,9 +1,10 @@
-use std::{collections::BTreeMap, sync::Arc};
+use std::collections::BTreeMap;
 
 use super::*;
 use libc::c_int;
 
-use voicevox_core::{OpenJtalk, StyleId, VoiceModel};
+use ndarray::ArrayView;
+use voicevox_core::{StyleId, VoiceModel, __internal::interp::PerformInference as _};
 
 macro_rules! ensure_initialized {
     ($synthesizer:expr $(,)?) => {
@@ -88,10 +89,10 @@ fn voice_model_set() -> &'static VoiceModelSet {
     &VOICE_MODEL_SET
 }
 
-static SYNTHESIZER: Lazy<Mutex<Option<voicevox_core::Synthesizer>>> =
+static SYNTHESIZER: Lazy<Mutex<Option<voicevox_core::Synthesizer<()>>>> =
     Lazy::new(|| Mutex::new(None));
 
-fn lock_synthesizer() -> MutexGuard<'static, Option<voicevox_core::Synthesizer>> {
+fn lock_synthesizer() -> MutexGuard<'static, Option<voicevox_core::Synthesizer<()>>> {
     SYNTHESIZER.lock().unwrap()
 }
 
@@ -106,7 +107,7 @@ fn set_message(message: &str) {
 pub extern "C" fn initialize(use_gpu: bool, cpu_num_threads: c_int, load_all_models: bool) -> bool {
     let result = RUNTIME.block_on(async {
         let synthesizer = voicevox_core::Synthesizer::new(
-            Arc::new(OpenJtalk::new_without_dic()),
+            (),
             &voicevox_core::InitializeOptions {
                 acceleration_mode: if use_gpu {
                     voicevox_core::AccelerationMode::Gpu
@@ -190,20 +191,21 @@ pub extern "C" fn supported_devices() -> *const c_char {
 }
 
 #[no_mangle]
-pub extern "C" fn yukarin_s_forward(
+pub unsafe extern "C" fn yukarin_s_forward(
     length: i64,
     phoneme_list: *mut i64,
     speaker_id: *mut i64,
     output: *mut f32,
 ) -> bool {
+    let length = length as usize;
     let synthesizer = &*lock_synthesizer();
-    let result = RUNTIME.block_on(ensure_initialized!(synthesizer).predict_duration(
-        unsafe { std::slice::from_raw_parts_mut(phoneme_list, length as usize) },
+    let result = ensure_initialized!(synthesizer).predict_duration(
+        ArrayView::from_shape_ptr((length,), phoneme_list).into_owned(),
         StyleId::new(unsafe { *speaker_id as u32 }),
-    ));
+    );
     match result {
         Ok(output_vec) => {
-            let output_slice = unsafe { std::slice::from_raw_parts_mut(output, length as usize) };
+            let output_slice = std::slice::from_raw_parts_mut(output, length);
             output_slice.clone_from_slice(&output_vec);
             true
         }
@@ -215,7 +217,7 @@ pub extern "C" fn yukarin_s_forward(
 }
 
 #[no_mangle]
-pub extern "C" fn yukarin_sa_forward(
+pub unsafe extern "C" fn yukarin_sa_forward(
     length: i64,
     vowel_phoneme_list: *mut i64,
     consonant_phoneme_list: *mut i64,
@@ -226,20 +228,20 @@ pub extern "C" fn yukarin_sa_forward(
     speaker_id: *mut i64,
     output: *mut f32,
 ) -> bool {
+    let length = length as usize;
     let synthesizer = &*lock_synthesizer();
-    let result = RUNTIME.block_on(ensure_initialized!(synthesizer).predict_intonation(
-        length as usize,
-        unsafe { std::slice::from_raw_parts(vowel_phoneme_list, length as usize) },
-        unsafe { std::slice::from_raw_parts(consonant_phoneme_list, length as usize) },
-        unsafe { std::slice::from_raw_parts(start_accent_list, length as usize) },
-        unsafe { std::slice::from_raw_parts(end_accent_list, length as usize) },
-        unsafe { std::slice::from_raw_parts(start_accent_phrase_list, length as usize) },
-        unsafe { std::slice::from_raw_parts(end_accent_phrase_list, length as usize) },
+    let result = ensure_initialized!(synthesizer).predict_intonation(
+        ArrayView::from_shape_ptr((length,), vowel_phoneme_list).into_owned(),
+        ArrayView::from_shape_ptr((length,), consonant_phoneme_list).into_owned(),
+        ArrayView::from_shape_ptr((length,), start_accent_list).into_owned(),
+        ArrayView::from_shape_ptr((length,), end_accent_list).into_owned(),
+        ArrayView::from_shape_ptr((length,), start_accent_phrase_list).into_owned(),
+        ArrayView::from_shape_ptr((length,), end_accent_phrase_list).into_owned(),
         StyleId::new(unsafe { *speaker_id as u32 }),
-    ));
+    );
     match result {
         Ok(output_vec) => {
-            let output_slice = unsafe { std::slice::from_raw_parts_mut(output, length as usize) };
+            let output_slice = std::slice::from_raw_parts_mut(output, length);
             output_slice.clone_from_slice(&output_vec);
             true
         }
@@ -251,7 +253,7 @@ pub extern "C" fn yukarin_sa_forward(
 }
 
 #[no_mangle]
-pub extern "C" fn decode_forward(
+pub unsafe extern "C" fn decode_forward(
     length: i64,
     phoneme_size: i64,
     f0: *mut f32,
@@ -262,16 +264,14 @@ pub extern "C" fn decode_forward(
     let length = length as usize;
     let phoneme_size = phoneme_size as usize;
     let synthesizer = &*lock_synthesizer();
-    let result = RUNTIME.block_on(ensure_initialized!(synthesizer).decode(
-        length,
-        phoneme_size,
-        unsafe { std::slice::from_raw_parts(f0, length) },
-        unsafe { std::slice::from_raw_parts(phoneme, phoneme_size * length) },
+    let result = ensure_initialized!(synthesizer).decode(
+        ArrayView::from_shape_ptr((length,), f0),
+        ArrayView::from_shape_ptr((length, phoneme_size), phoneme),
         StyleId::new(unsafe { *speaker_id as u32 }),
-    ));
+    );
     match result {
         Ok(output_vec) => {
-            let output_slice = unsafe { std::slice::from_raw_parts_mut(output, length * 256) };
+            let output_slice = std::slice::from_raw_parts_mut(output, length * 256);
             output_slice.clone_from_slice(&output_vec);
             true
         }
