+++ crates/voicevox_core_c_api/src/lib.rs
@@ -319,7 +319,7 @@ pub extern "C" fn voicevox_voice_model_delete(model: Box<VoicevoxVoiceModel>) {
 /// <b>構築</b>(_construction_)は ::voicevox_synthesizer_new で行い、<b>破棄</b>(_destruction_)は ::voicevox_synthesizer_delete で行う。
 #[derive(Getters)]
 pub struct VoicevoxSynthesizer {
-    synthesizer: Synthesizer,
+    synthesizer: Synthesizer<Arc<OpenJtalk>>,
 }
 
 /// ::VoicevoxSynthesizer を<b>構築</b>(_construct_)する。
