+++ crates/voicevox_core_java_api/src/synthesizer.rs
@@ -9,6 +9,7 @@ use jni::{
     JNIEnv,
 };
 use std::sync::Arc;
+use voicevox_core::OpenJtalk;
 
 #[no_mangle]
 unsafe extern "system" fn Java_jp_hiroshiba_voicevoxcore_Synthesizer_rsNew<'local>(
@@ -64,7 +65,9 @@ unsafe extern "system" fn Java_jp_hiroshiba_voicevoxcore_Synthesizer_rsIsGpuMode
 ) -> jboolean {
     throw_if_err(env, false, |env| {
         let internal = env
-            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer>>(&this, "handle")?
+            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer<Arc<OpenJtalk>>>>(
+                &this, "handle",
+            )?
             .clone();
 
         Ok(internal.is_gpu_mode())
@@ -78,7 +81,9 @@ unsafe extern "system" fn Java_jp_hiroshiba_voicevoxcore_Synthesizer_rsGetMetasJ
 ) -> jobject {
     throw_if_err(env, std::ptr::null_mut(), |env| {
         let internal = env
-            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer>>(&this, "handle")?
+            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer<Arc<OpenJtalk>>>>(
+                &this, "handle",
+            )?
             .clone();
 
         let metas_json = serde_json::to_string(&internal.metas()).expect("should not fail");
@@ -100,7 +105,9 @@ unsafe extern "system" fn Java_jp_hiroshiba_voicevoxcore_Synthesizer_rsLoadVoice
             .get_rust_field::<_, _, Arc<voicevox_core::VoiceModel>>(&model, "handle")?
             .clone();
         let internal = env
-            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer>>(&this, "handle")?
+            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer<Arc<OpenJtalk>>>>(
+                &this, "handle",
+            )?
             .clone();
         RUNTIME.block_on(internal.load_voice_model(&model))?;
         Ok(())
@@ -117,7 +124,9 @@ unsafe extern "system" fn Java_jp_hiroshiba_voicevoxcore_Synthesizer_rsUnloadVoi
         let model_id: String = env.get_string(&model_id)?.into();
 
         let internal = env
-            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer>>(&this, "handle")?
+            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer<Arc<OpenJtalk>>>>(
+                &this, "handle",
+            )?
             .clone();
 
         internal.unload_voice_model(&voicevox_core::VoiceModelId::new(model_id))?;
@@ -138,7 +147,9 @@ unsafe extern "system" fn Java_jp_hiroshiba_voicevoxcore_Synthesizer_rsIsLoadedV
         let model_id: String = env.get_string(&model_id)?.into();
 
         let internal = env
-            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer>>(&this, "handle")?
+            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer<Arc<OpenJtalk>>>>(
+                &this, "handle",
+            )?
             .clone();
 
         let is_loaded = internal.is_loaded_voice_model(&voicevox_core::VoiceModelId::new(model_id));
@@ -162,7 +173,9 @@ unsafe extern "system" fn Java_jp_hiroshiba_voicevoxcore_Synthesizer_rsAudioQuer
         let style_id = style_id as u32;
 
         let internal = env
-            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer>>(&this, "handle")?
+            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer<Arc<OpenJtalk>>>>(
+                &this, "handle",
+            )?
             .clone();
 
         let audio_query = RUNTIME.block_on(
@@ -189,7 +202,9 @@ unsafe extern "system" fn Java_jp_hiroshiba_voicevoxcore_Synthesizer_rsAudioQuer
         let style_id = style_id as u32;
 
         let internal = env
-            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer>>(&this, "handle")?
+            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer<Arc<OpenJtalk>>>>(
+                &this, "handle",
+            )?
             .clone();
 
         let audio_query =
@@ -217,7 +232,9 @@ unsafe extern "system" fn Java_jp_hiroshiba_voicevoxcore_Synthesizer_rsAccentPhr
         let style_id = style_id as u32;
 
         let internal = env
-            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer>>(&this, "handle")?
+            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer<Arc<OpenJtalk>>>>(
+                &this, "handle",
+            )?
             .clone();
 
         let accent_phrases = RUNTIME.block_on(
@@ -244,7 +261,9 @@ unsafe extern "system" fn Java_jp_hiroshiba_voicevoxcore_Synthesizer_rsAccentPhr
         let style_id = style_id as u32;
 
         let internal = env
-            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer>>(&this, "handle")?
+            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer<Arc<OpenJtalk>>>>(
+                &this, "handle",
+            )?
             .clone();
 
         let accent_phrases = RUNTIME.block_on(
@@ -273,7 +292,9 @@ unsafe extern "system" fn Java_jp_hiroshiba_voicevoxcore_Synthesizer_rsReplaceMo
         let style_id = style_id as u32;
 
         let internal = env
-            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer>>(&this, "handle")?
+            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer<Arc<OpenJtalk>>>>(
+                &this, "handle",
+            )?
             .clone();
 
         let replaced_accent_phrases = RUNTIME.block_on(
@@ -303,7 +324,9 @@ unsafe extern "system" fn Java_jp_hiroshiba_voicevoxcore_Synthesizer_rsReplacePh
         let style_id = style_id as u32;
 
         let internal = env
-            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer>>(&this, "handle")?
+            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer<Arc<OpenJtalk>>>>(
+                &this, "handle",
+            )?
             .clone();
 
         let replaced_accent_phrases = {
@@ -334,7 +357,9 @@ unsafe extern "system" fn Java_jp_hiroshiba_voicevoxcore_Synthesizer_rsReplaceMo
         let style_id = style_id as u32;
 
         let internal = env
-            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer>>(&this, "handle")?
+            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer<Arc<OpenJtalk>>>>(
+                &this, "handle",
+            )?
             .clone();
 
         let replaced_accent_phrases = RUNTIME.block_on(
@@ -363,7 +388,9 @@ unsafe extern "system" fn Java_jp_hiroshiba_voicevoxcore_Synthesizer_rsSynthesis
         let style_id = style_id as u32;
 
         let internal = env
-            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer>>(&this, "handle")?
+            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer<Arc<OpenJtalk>>>>(
+                &this, "handle",
+            )?
             .clone();
 
         let wave = {
@@ -397,7 +424,9 @@ unsafe extern "system" fn Java_jp_hiroshiba_voicevoxcore_Synthesizer_rsTtsFromKa
         let style_id = style_id as u32;
 
         let internal = env
-            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer>>(&this, "handle")?
+            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer<Arc<OpenJtalk>>>>(
+                &this, "handle",
+            )?
             .clone();
 
         let wave = {
@@ -431,7 +460,9 @@ unsafe extern "system" fn Java_jp_hiroshiba_voicevoxcore_Synthesizer_rsTts<'loca
         let style_id = style_id as u32;
 
         let internal = env
-            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer>>(&this, "handle")?
+            .get_rust_field::<_, _, Arc<voicevox_core::Synthesizer<Arc<OpenJtalk>>>>(
+                &this, "handle",
+            )?
             .clone();
 
         let wave = {
