+++ crates/voicevox_core_python_api/python/voicevox_core/__init__.py
@@ -21,10 +21,10 @@ from ._rust import (  # noqa: F401
     InferenceFailedError,
     InvalidModelDataError,
     InvalidWordError,
+    LoadOpenjtalkSystemDicError,
     LoadUserDictError,
     ModelAlreadyLoadedError,
     ModelNotFoundError,
-    NotLoadedOpenjtalkDictError,
     OpenJtalk,
     OpenZipFileError,
     ParseKanaError,
@@ -52,11 +52,11 @@ __all__ = [
     "InferenceFailedError",
     "InvalidModelDataError",
     "InvalidWordError",
+    "LoadOpenjtalkSystemDicError",
     "LoadUserDictError",
     "ModelAlreadyLoadedError",
     "ModelNotFoundError",
     "Mora",
-    "NotLoadedOpenjtalkDictError",
     "OpenJtalk",
     "OpenZipFileError",
     "ParseKanaError",
